Changed hidden paste textarea to fixed to prevent scrolling when pasting
When pasting into the notebook, content is written to a hidden textarea. The paste capturing script focuses on this textarea, which causes the browser to scroll to it. On the demo, the textarea is already in view, so you don't see the change. Changing the textarea to fixed instead of absolute keeps it in view so the browser doesn't need to scroll.